+++ docs/release-notes.md
@@ -1,5 +1,6 @@
 # New features and Changes
 - Shorten long anime titles in progress bars
+- Hover on folder button shows folder location
 
 # Bug fixes
 - Fix failed to fetch direct download links notification on cancelling direct download links retrieval